add timezones

+++ e3f2s/simulator/multiple_runs/multiple_runs.py
@@ -34,21 +34,20 @@ def multiple_runs(sim_run_conf, sim_general_conf, sim_scenario_conf_grid, sim_sc
 
 	with mp.Pool(n_cores) as pool:
 
-		# if not os.path.exists(os.path.join(results_path, "city_obj.pickle")):
-		city_obj = City(city, sim_run_conf["data_source_id"], sim_general_conf)
-		pickle.dump(
-			city_obj,
-			open(os.path.join(results_path, "city_obj.pickle"), "wb")
-		)
-		city_obj.grid_matrix.to_pickle(
-			os.path.join(results_path, "grid_matrix.pickle")
-		)
-		pd.DataFrame(city_obj.neighbors_dict).to_pickle(
-			os.path.join(results_path, "neighbors_dict.pickle")
-		)
-
-		# else:
-		# 	city_obj = pickle.Unpickler(open(os.path.join(results_path, "city_obj.pickle"), "rb")).load()
+		if not os.path.exists(os.path.join(results_path, "city_obj.pickle")):
+			city_obj = City(city, sim_run_conf["data_source_id"], sim_general_conf)
+			pickle.dump(
+				city_obj,
+				open(os.path.join(results_path, "city_obj.pickle"), "wb")
+			)
+			city_obj.grid_matrix.to_pickle(
+				os.path.join(results_path, "grid_matrix.pickle")
+			)
+			pd.DataFrame(city_obj.neighbors_dict).to_pickle(
+				os.path.join(results_path, "neighbors_dict.pickle")
+			)
+		else:
+			city_obj = pickle.Unpickler(open(os.path.join(results_path, "city_obj.pickle"), "rb")).load()
 
 		sim_conf_grid = EFFCS_SimConfGrid(sim_scenario_conf_grid)
 
@@ -96,27 +95,3 @@ def multiple_runs(sim_run_conf, sim_general_conf, sim_scenario_conf_grid, sim_sc
 	sim_stats_df.to_pickle(os.path.join(results_path, "sim_stats.pickle"))
 	pd.Series(sim_general_conf).to_pickle(os.path.join(results_path, "sim_general_conf.pickle"))
 	pd.Series(sim_scenario_conf_grid).to_pickle(os.path.join(results_path, "sim_scenario_conf_grid.pickle"))
-
-	# plotter = EFFCS_MultipleRunsPlotter(
-	# 	city, sim_scenario_name, sim_general_conf, sim_scenario_conf_grid,
-	# 	"alpha", "fraction_unsatisfied", "beta"
-	# )
-	# plotter.plot_x_y_param()
-	#
-	# plotter = EFFCS_MultipleRunsPlotter(
-	# 	city, sim_scenario_name, sim_general_conf, sim_scenario_conf_grid,
-	# 	"alpha", "fraction_not_same_zone_trips", "beta"
-	# )
-	# plotter.plot_x_y_param()
-	#
-	# plotter = EFFCS_MultipleRunsPlotter(
-	# 	city, sim_scenario_name, sim_general_conf, sim_scenario_conf_grid,
-	# 	"alpha", "fraction_no_close_cars", "beta"
-	# )
-	# plotter.plot_x_y_param()
-	#
-	# plotter = EFFCS_MultipleRunsPlotter(
-	# 	city, sim_scenario_name, sim_general_conf, sim_scenario_conf_grid,
-	# 	"alpha", "fraction_not_enough_energy", "beta"
-	# )
-	# plotter.plot_x_y_param()
